Fixed logf import

+++ controllers/multiclusterhub_controller.go
@@ -481,7 +481,7 @@ func (r *MultiClusterHubReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 		return ctrl.Result{RequeueAfter: resyncPeriod}, nil
 	}
 
-	log.Log.Info("Reconcile completed. Requeuing after " + utils.ShortRefreshInterval.String())
+	logf.Log.Info("Reconcile completed. Requeuing after " + utils.ShortRefreshInterval.String())
 	return ctrl.Result{RequeueAfter: utils.ShortRefreshInterval}, nil
 }
